SKC-008
- added "docker-compose-integration-tests" project to
"springboot-testing-tips". It is an adapted maven version, with the
latest spring boot and junit5. The original can be found the following
article: https://blog.codecentric.de/en/2017/03/writing-integration-tests-docker-compose-junit/

The project shows how to write the test for dynamo-db, starting it as a
docker container with docker-compose script. The project can be adapted
for any kind of database or service as docker image. It should be
possible to handle any number of docker-images in one docker-compose
script, within one test, without any manual docker commands preparation.